requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.1
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.8.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pytest==8.4.1
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
zipp==3.23.0
Flask-JWT-Extended==4.5.2
flask-marshmallow==1.3.0
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2